refactor(transaction): TransactionProvider에서 fetchTransactions 수행하도록 수정
- 일렉트럼 스트림 종료 안되는 오류 수정
