+++ lib/screens/wallet_detail/wallet_detail_screen.dart
@@ -202,7 +202,6 @@ class _WalletDetailScreenState extends State<WalletDetailScreen> {
                               }
                               viewModel.walletProvider
                                   ?.initWallet(targetId: widget.id);
-                              viewModel.fetchTransactions();
                             } finally {
                               _isPullToRefreshing = false;
                             }
